+++ R/util-print-cvm-tbl.R
@@ -6,6 +6,7 @@
 print.cvm_tbl <- function(x, ...) {
   src <- attr(x, "source", exact = TRUE)
   fetched <- attr(x, "fetched_at", exact = TRUE)
+  grp <- attr(x, "group", exact = TRUE)
   ds <- attr(x, "dataset", exact = TRUE)
   tbl <- attr(x, "table", exact = TRUE)
   cli::cli_inform(c(
@@ -14,6 +15,7 @@ print.cvm_tbl <- function(x, ...) {
       "{.field fetched_at}: {.val {fetched}}"
     ),
     "i" = paste0(
+      "{.field group}: {.val {grp}}  |  ",
       "{.field dataset}: {.val {ds}}  |  ",
       "{.field table}: {.val {tbl}}"
     )